Make the "EntityChangeSet" available in PostUpdate
